# Fix race condition for trial number computation.