## Motivation

Fixes https://github.com/optuna/optuna/issues/1488.

## Description of the changes

Allows `count_past_trials` to circumvent repeatable read isolation level. See the comments in the code for details.

## Note

Verified the changes with MySQL, by inspection, by checking for unique trial numbers after running a distributed optimization. The bug is otherwise reproducible, simply running a lot of trials with.

```sql
SELECT COUNT(DISTINCT(number)) FROM trials;
```

```python
import optuna
from optuna.samplers import TPESampler

def objective(trial, n_params):
    return sum(trial.suggest_float(f"x{i}", 0.0, 1.0) for i in range(n_params))

if __name__ == "__main__":
    n_params = 10
    n_trials = 200

    database = "duplicatenumber1488"
    storage = "mysql://root@localhost/duplicatenumber1488"
    study = optuna.create_study(sampler=TPESampler(), study_name=database, storage=storage, load_if_exists=True)

    study.optimize(lambda trial: objective(trial, n_params), n_trials=n_trials)
```

**About isolation levels in sqlalchemy for different dialect** 
https://docs.sqlalchemy.org/en/13/core/connections.html#sqlalchemy.engine.Connection.execution_options.params.isolation_level.