@c-bata 2020-07-07
"Read Uncommitted" is lower isolation level than "Repeatable Read". It looks that this PR depends on "Dirty Reads", right?

> ![ScreenShot 2020-07-07 18 17 57](https://user-images.githubusercontent.com/5564044/86759532-8e5c1380-c07f-11ea-8526-658ae3486585.png)
> https://en.wikipedia.org/wiki/Isolation_(database_systems)

Like this table said, "Dirty reads" **may occure** when using "Read Uncommitted". So it sounds that this fixes is not safe.