@hvy 2020-07-07
<details><summary>Code</summary>
<p>

```python
import argparse
import math
import time

import optuna
from optuna.samplers import TPESampler
import sqlalchemy


class Profile:
    def __enter__(self):
        self.start = time.time()
        return self

    def __exit__(self, exc_type, exc_val, exc_tb):
        self.end = time.time()

    def get(self):
        return self.end - self.start


def build_objective_fun(n_param):
    def objective(trial):
        return sum(
            [
                math.sin(trial.suggest_uniform("param-{}".format(i), 0, math.pi * 2))
                for i in range(n_param)
            ]
        )

    return objective


if __name__ == "__main__":
    parser = argparse.ArgumentParser()
    parser.add_argument("mysql_user", type=str)
    parser.add_argument("mysql_host", type=str)
    args = parser.parse_args()

    storage_str = "mysql+pymysql://{}@{}/".format(args.mysql_user, args.mysql_host)

    optuna.logging.set_verbosity(optuna.logging.CRITICAL)

    print(f"| #params | #trials | time(sec) |")
    print(f"| ------- | ------- | --------- |")

    for n_param in [1, 2, 4, 8, 16, 32]:
        for n_trial in [1, 10, 100, 1000]:
            engine = sqlalchemy.create_engine(storage_str)
            conn = engine.connect()
            conn.execute("commit")
            database_str = "profile_storage_t{}_p{}".format(n_trial, n_param)
            try:
                conn.execute("drop database {}".format(database_str))
            except Exception:
                pass
            conn.execute("create database {}".format(database_str))
            conn.close()

            storage = optuna.storages.get_storage(storage_str + database_str)
            study = optuna.create_study(storage=storage, sampler=TPESampler())

            with Profile() as prof:
                study.optimize(
                    build_objective_fun(n_param), n_trials=n_trial, gc_after_trial=False,
                )

            print(f"| {n_param} | {n_trial} | {prof.get():.2f} |")
```

</p>
</details>

| #params | #trials | PR(sec) | `master`(sec) |
| ------- | ------- | ---- | ------------- |
| 1 | 1 | 0.02 | 0.01 |
| 1 | 10 | 0.12 | 0.11 |
| 1 | 100 | 1.60 | 1.18 |
| 1 | 1000 | 35.97 | 16.84 |
| 2 | 1 | 0.02 | 0.02 |
| 2 | 10 | 0.13 | 0.11 |
| 2 | 100 | 1.53 | 1.25 |
| 2 | 1000 | 38.03 | 20.17 |
| 4 | 1 | 0.02 | 0.02 |
| 4 | 10 | 0.14 | 0.12 |
| 4 | 100 | 1.83 | 1.53 |
| 4 | 1000 | 42.60 | 26.23 |
| 8 | 1 | 0.04 | 0.03 |
| 8 | 10 | 0.16 | 0.15 |
| 8 | 100 | 2.36 | 2.04 |
| 8 | 1000 | 52.53 | 36.11 |
| 16 | 1 | 0.08 | 0.06 |
| 16 | 10 | 0.20 | 0.20 |
| 16 | 100 | 3.21 | 2.98 |
| 16 | 1000 | 71.93 | 54.78 |
| 32 | 1 | 0.11 | 0.11 |
| 32 | 10 | 0.30 | 0.29 |
| 32 | 100 | 4.95 | 4.64 |
| 32 | 1000 | 110.76 | 93.52 |