@ytsmiling 2020-07-09
It's highly unlikely, but if trial_id is not assigned sequentially, the `count_past_trial` method can fail (duplicated number can be assigned). I'm okay about leaving the method as is, but if you'd like to change the implementation, I'll re-review this PR.